森下 11/7 画面に入荷数、受注数、出荷数、注残を表示しました。 数値をLongに統一しました。

+++ BeerGame/src/java/jp/co/isken/beerGame/entity/TradeTransaction.java
@@ -19,45 +19,45 @@ public class TradeTransaction extends jp.co.isken.beerGame.entity.base.BaseTrade
 
 	private static final long serialVersionUID = 1L;
 
-	public static Map<Integer, Integer> getRemainAmount(String gameName,
+	public static Map<Long, Long> getRemainAmount(String gameName,
 			String roleName) {
 		Game game = Game.getGameByName(gameName);
 		Role role = game.getRole(roleName);
 		//FIXME これでいいのか？　現在は仮で受注の最終週を取得している　2009/10/31
-		return getRemainList(role.getWeek(TransactionType.受注.name()).intValue(), role);
+		return getRemainList(role.getWeek(TransactionType.受注.name()), role);
 	}
 
-	private static Map<Integer, Integer> getRemainList(int week, Role role) {
-		Map<Integer, Integer> ret = new HashMap<Integer, Integer>();
-		for(int i = 1 ; i <= week ; i++){
+	private static Map<Long, Long> getRemainList(Long week, Role role) {
+		Map<Long, Long> ret = new HashMap<Long, Long>();
+		for(Long i = 1L ; i <= week ; i++){
 			ret.put(i, calcAmountRemain(i, role));
 		}
 		return ret;
 	}
 
-	public static int calcAmountRemain(int week, Role role) {
+	public static Long calcAmountRemain(Long week, Role role) {
 		return calcAmount(week, role, "受注") - calcAmount(week, role, "出荷");	}
 
-	public static Map<Integer, Integer> getStockAmount(String gameName, String roleName) {
+	public static Map<Long, Long> getStockAmount(String gameName, String roleName) {
 		Game game = Game.getGameByName(gameName);
 		Role role = game.getRole(roleName);
 		//FIXME これでいいのか？　現在は仮で受注の最終週を取得している　2009/10/31
-		return getStockList(role.getWeek(TransactionType.入荷.name()).intValue()-1, role);
+		return getStockList(role.getWeek(TransactionType.入荷.name())-1, role);
 	}
 
-	public static Map<Integer, Integer> getStockList(int week, Role role) {
-		Map<Integer, Integer> ret = new HashMap<Integer, Integer>();
-		for(int i = 1 ; i <= week ; i++){
+	public static Map<Long, Long> getStockList(Long week, Role role) {
+		Map<Long, Long> ret = new HashMap<Long, Long>();
+		for(Long i = 1L ; i <= week ; i++){
 			ret.put(i, calcAmountStock(i, role));
 		}
 		return ret;
 	}
 
-	public static int calcAmountStock(int week, Role role) {
+	public static Long calcAmountStock(Long week, Role role) {
 		return calcAmount(week, role, "入荷") - calcAmount(week, role, "出荷");
 	}
 
-	public static int calcAmount(int week, Role role, String typeName) {
+	public static Long calcAmount(Long week, Role role, String typeName) {
 		Extractor ext = new Extractor(TradeTransaction.class);
 		ext.addExtractValue(new ExtractValue(
 				"sum", new Sum(new Property(TradeTransaction.AMOUNT))));
@@ -65,6 +65,6 @@ public class TradeTransaction extends jp.co.isken.beerGame.entity.base.BaseTrade
 		ext.add(Condition.le(new Property(TradeTransaction.WEEK), week));
 		ext.add(Condition.eq(new Property(TradeTransaction.TRANSACTION_TYPE), typeName));
 		List<Map<String, Long>> list = BasicService.getService().findByExtractor(ext);
-		return list.get(0).get("sum").intValue();	
+		return list.get(0).get("sum");	
 	}
 }
